format

+++ interfaces/fragment/gart_fragment.h
@@ -268,8 +268,10 @@ class GartFragment {
       auto src_id = vertex_name_id_map.find(src_name)->second;
       auto dst_id = vertex_name_id_map.find(dst_name)->second;
 
-      vertex2edge_map.emplace(std::make_pair(src_id, dst_id), e_label_id - vertex_label_num_);
-      edge2vertex_map.emplace(e_label_id - vertex_label_num_, std::make_pair(src_id, dst_id));
+      vertex2edge_map.emplace(std::make_pair(src_id, dst_id),
+                              e_label_id - vertex_label_num_);
+      edge2vertex_map.emplace(e_label_id - vertex_label_num_,
+                              std::make_pair(src_id, dst_id));
 
       edge_label_num_++;
     }
@@ -604,7 +606,7 @@ class GartFragment {
       computeVertexNum();
     }
     size_t total_num = 0;
-    for (size_t idx = 0; idx < vertex_label_num_; ++idx) {
+    for (int idx = 0; idx < vertex_label_num_; ++idx) {
       total_num += ivnums_[idx];
     }
     return total_num;
